- upravil jsem dialogové okno --> pokud je výběr pomoci katuzID "checked" a ID vloženo, je ukladáno temporální CSV soubor s ID číslem (které se po načtení vrstvy vymaže).
PS: Po načtení vrstvy, přechazí režim (vrstva dle ID) zpátky do režimu (vrstva dle středu okna).

+++ data_sources/LPIS_view/options_dialog.py
@@ -39,23 +39,21 @@ class OptionsDialog(QtWidgets.QDialog, FORM_CLASS):
         """Constructor."""
         super(OptionsDialog, self).__init__(parent)
         self.setupUi(self)
-        self.pushButtonConfirm.clicked.connect(self.katuzid_save)
-        self.pushButtonUnconfirm.clicked.connect(self.katuz_clean)
+        self.buttonBox.accepted.connect(self.accepted)
 
-    def katuzid_save(self):
-        katuzidstored = self.lineEdit.text()
-        current_dir_to_id = os.path.join(os.path.dirname(os.path.abspath(__file__)))
-        path_to_id = os.path.join(current_dir_to_id, 'data', 'temporar.csv')
-        with open(path_to_id, "w") as writer:
+
+    def accepted(self):
+        if self.radioButtonId.isChecked():
+            katuzidstored = self.lineEdit.text()
+            current_dir_to_id = os.path.join(os.path.dirname(os.path.abspath(__file__)))
+            path_to_id = os.path.join(current_dir_to_id, 'data', 'temporar.csv')
+            with open(path_to_id, "w") as writer:
                 writer.write(katuzidstored)
                 writer.close()
-        if not self.lineEdit.text():
+            if not self.lineEdit.text():
+                os.remove(path_to_id)
+        else:
             os.remove(path_to_id)
-
-    def katuz_clean(self):
-        current_dir_to_id = os.path.join(os.path.dirname(os.path.abspath(__file__)))
-        path_to_id = os.path.join(current_dir_to_id, 'data', 'temporar.csv')
-        os.remove(path_to_id)
-
+            return
